overhaul Synchronize Maven Artifacts task
Drop classpath messing, selectively discard index jobs:

Since we do refreshLocal() on the updated maven resource file, there
should be no more need to mess with the project's classpath. This seems
to have caused build failures, triggering the workaround to close+open
the project for a rebuild. However, messing with classpath might have
been necessary as workaround with Eclipse 2021-12 because of some
indexer changes there, and seems to be obsolete with Eclipse 2022-06.

Since we compare the existing maven resource for whether to be updated,
we can stop the indexer jobs for updated or deleted resources only.

Also, re-enable the indexer before creating workspace change events,
might help to avoid more close+open rebuilds.
And for the file operations we use the raw file system API, for the
workspace events to be generated by the refresh call only. Important
here is to surround with the workspace checkpoint API call, though.

Plus, avoid a non-final variable static to the JVM.

Change-Id: I28787e5ddfa000a80995fca879631fd35ff7abb2