ClasspathUpdater: preserve existing entries
Refreshing the classpath does not mean to remove entries that are not
created per the PDE level. Instead, just keep them unchanged.
Also, for existing SOURCE entries, stick to the original path notation,
in order to avoid adding a trailing slash, which might be provided when
recreating classpath from scratch.

Change-Id: Ibd9f356b541336b7b85bf8a95c398894988f5849